requirements.txt
argcomplete==1.11.1
boto3==1.12.34
botocore==1.15.34
certifi==2019.11.28
cfn-flip==1.2.2
chardet==3.0.4
click==7.1.1
docutils==0.15.2
durationpy==0.5
Flask==1.1.1
future==0.18.2
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
kappa==0.6.0
MarkupSafe==1.1.1
Naked==0.1.31
pip-tools==4.5.1
placebo==0.9.0
pycrypto==2.6.1
pymongo==3.10.1
python-dateutil==2.6.1
python-slugify==4.0.0
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
shellescape==3.8.1
six==1.14.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.44.1
troposphere==2.6.0
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
